refactor: drop doobie-quill integration
* fewer dependencies are better

+++ src/main/scala/net/yoshinorin/qualtet/domains/models/authors/AuthorRepository.scala
@@ -8,9 +8,9 @@ trait AuthorRepository {
    * create a authorName
    *
    * @param data Instance of Author
-   * @return dummy long id (Doobie return Long)
+   * @return dummy long id (Doobie return Int)
    */
-  def upsert(data: Author): ConnectionIO[Long]
+  def upsert(data: Author): ConnectionIO[Int]
 
   /**
    * get all Author
