+++ src/main/scala/net/yoshinorin/qualtet/domains/models/robots/RobotsRepository.scala
@@ -9,9 +9,9 @@ trait RobotsRepository {
    * create a robots (for meta)
    *
    * @param data Instance of Robots (for meta)
-   * @return dummy long id (Doobie return Long)
+   * @return dummy long id (Doobie return Int)
    */
-  def upsert(data: Robots): ConnectionIO[Long]
+  def upsert(data: Robots): ConnectionIO[Int]
 
   /**
    * find a robots by ContentId
